pyproject.toml
[project]
name = "wp-theme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
    "pyyaml>=6.0.0",
    "deepagents>=0.4.12",
    "langchain-openai>=1.1.12",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
